ml-lab/requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
opencv-python==4.10.0.84
Pillow==11.0.0
PyWavelets==1.7.0
PyYAML==6.0.2
joblib==1.4.2
tqdm==4.67.1
matplotlib==3.9.2
seaborn==0.13.2
tabulate==0.9.0
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.12
